third_party/WebKit/LayoutTests/http/tests/security/cors-rfc1918/addressspace-serviceworker-basic.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./resources/addressspace-test.js"></script>
<script>
    window.onload = function () {
        addressSpaceTest("http://localhost:8000", "serviceworker", "local");
        addressSpaceTest("http://127.0.0.1:8000", "serviceworker", "local");
        addressSpaceTest("https://example.test:8443", "serviceworker", "private");
    };
</script>
